Move all cosmological tools to cosmology.py
