+++ healvis/tests/test_pspec.py
@@ -2,7 +2,7 @@ import numpy as np
 from astropy.time import Time
 import nose.tools as nt
 
-from healvis import observatory, sky_model, beam_model, utils, cosmo_funcs
+from healvis import observatory, sky_model, beam_model, utils, cosmology
 
 latitude = -30.7215277777
 longitude = 21.4283055554
@@ -44,7 +44,7 @@ def test_pspec_amp():
     visibs, times, bls = obs.make_visibilities(sky, Nprocs=3)
 
     vis_jy = visibs[:, 0, :]  # (Nblts, Nskies, Nfreqs)
-    vis_Ksr = cosmo_funcs.jy2Tstr(freqs) * vis_jy
+    vis_Ksr = cosmology.jy2Tsr(freqs) * vis_jy
 
     _vis = np.fft.ifft(vis_Ksr, axis=1)
     Nkpar = Nfreqs / 2
@@ -57,14 +57,14 @@ def test_pspec_amp():
     beam_sq_int = beam_sq_int[0]    # Constant across frequencies
 
     Bandwidth = freqs[-1] - freqs[0]
-    scalar = cosmo_funcs.X2Y(sky.Z_array[sky.ref_chan]) * (Bandwidth / beam_sq_int)
+    scalar = cosmology.X2Y(sky.Z_array[sky.ref_chan]) * (Bandwidth / beam_sq_int)
 
     dspec_I = np.mean(dspec_instr * scalar, axis=0)
 
     # Theoretical pspec amp
     dnu = np.diff(freqs)[0] / 1e6  # MHz
     Z_sel = sky.Z_array[sky.ref_chan]
-    amp_theor = skysig**2 * utils.comoving_voxel_volume(Z_sel, dnu, sky.pix_area_sr)
+    amp_theor = skysig**2 * cosmology.comoving_voxel_volume(Z_sel, dnu, sky.pix_area_sr)
 
     tolerance = (amp_theor / float(Ntimes))   # assuming independent fields
     print(amp_theor, np.mean(dspec_I))
